fix: race condition for device error propagation fix: NO_SENSE errortype when it should be backwards fix: unified simple problems
